+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/utils/ExternalCommand.scala
@@ -1,6 +1,7 @@
 package io.joern.x2cpg.utils
 
 import java.io.File
+import java.nio.file.{Path, Paths}
 import java.util.concurrent.ConcurrentLinkedQueue
 import scala.sys.process.{Process, ProcessLogger}
 import scala.util.{Failure, Success, Try}
@@ -53,4 +54,26 @@ trait ExternalCommand {
   }
 }
 
-object ExternalCommand extends ExternalCommand
+object ExternalCommand extends ExternalCommand {
+
+  /** Finds the absolute path to the executable directory (e.g. `/path/to/javasrc2cpg/bin`).
+   * Based on the package path of a loaded classfile based on some (potentially flakey?) filename heuristics.
+   * Context: we want to be able to invoke the x2cpg frontends from any directory, not just their install directory,
+   * and then invoke other executables, like astgen, php-parser et al. 
+   * */
+  def executableDir(packagePath: String): Path = {
+    val dir        = packagePath
+    val indexOfLib = dir.lastIndexOf("lib")
+    val fixedDir = if (indexOfLib != -1) {
+      new java.io.File(dir.substring("file:".length, indexOfLib)).toString
+    } else {
+      val indexOfTarget = dir.lastIndexOf("target")
+      if (indexOfTarget != -1) {
+        new java.io.File(dir.substring("file:".length, indexOfTarget)).toString
+      } else {
+        "."
+      }
+    }
+    Paths.get(fixedDir, "/bin/").toAbsolutePath
+  }
+}
